refactored the Opiton, Options, and Namespaces classes.
Options and Namespaces classes now proxy a hash instead of
subclassing one.  Better pattern, more extendable.

The Option class expects the 3 params it needs already parsed.  This
will allows for things to inject rules and other args when building
new options.  Had to go rework the setup of most of the Option
tests to make them send 3 args always.

The Options class now parses the args directly.  The `args`
singleton method parses the args and returns them as an array
in the order the Option constructor expects them in (reverse of
previous order).  Had to rework Options and AssertMacros to
handle the args in this new reversed order.  Going to inject a
:value rule here when mass handling :value options.

Also, added the concept of an `option_type` to the options class.
This will allow for specialized handling of options for different
types.  Going to be used in mass handling :value options.