reworked the Option class constructor signature
Moved all fuzzy args out - now requires the 3 params to be passed
to the constructor.  The fuzzy args handling is still in effect,
its just used as a 3rd party service (the Options handler creates
its options using it).

This also reverses the output of the singleton `args method.

The reason I did this is that it is easier to have 3rd party things
inject or add to the default output from the args singleton this way.