move the namespace method_missing handling to the data class
Since all of the logic for handling this is based on the data, do the
handling there.

Reworked it to be more readable and encapsulate some decision logic
behind named local methods.  This closes #57.

Updated the handling to raise NoMethodError if trying to write a value
to a namespace.  Before, an option would have been written that could
never be read.  The net effect for the user was "it ignored it".  Both
this net effect and needlessly adding an option are not desirable.  Plus,
this makes nested namespace method behavior more congruent with root level
namespace method behavior (now both raise NoMethodError when writing to a
namespace).