Handler defining same-named options and/or namespaces
No matter the permutation (opt-over-ns or ns-over-opt), when adding
an option, remove any namespaces with the same name that happen to
exist and vice-versa when adding a namespace.

This ensures that only one option or namespace exists for a given
name at any time - the last one to be defined/added.  This also
removes the subjective task of determining which takes precedent
when a named thing is requested and there is both an option and a
namespace with the requested name.

The net effect is an intuitive experience for the user: the last thing
added with the name {name} will be return when requesting {name}.

Closes #58.