reworked the name tracking for a Namespace
The name value is really only _needed_ for the custom inspect on
the ns class.

Moved it off of the data class and stored as an obfuscated instance
var (to not pollute the method-space) on the ns.

There was no need for the data class to track it - and it could have
been dangerous because of collisions with local vars called `name`.