reworked how namespaces and options are added to the data collections
Moved the creating of the opt/ns instances into the data class.

This simplified the `add` methods on the collection classes.  This also
removed a layer of *args (pass-thru) handling between the data class
and the collection classes, which ultimately simplified things and
removed some unecessary abstraction.

This allows for a clearer path for the data class to change how
options and namespaces are built.