build Array type-cast options using `Array.new([*value])`
This is more robust type-casting than the previous `Array.new(value)`
method.  Previously, if you set an Array option equal to some other
value, it would attempt to convert the value to an integer and then
build the array composed of that many nil values.  This default
`Array.new` behavior isn't very helpful or intuitive.

With this update, you can set a single value and the option will
type-cast the value to a single element array with the value. This
is a much more intuitive behavior.