Fixnum coercion handling
Not previously handled so would error if given anything but an
integer.  That was inconsistant though because `Integer` type class
would convert strings just fine.  Why should the user be penalized
because they chose to use a `Fixnum` type class instead of `Integer`.
They shouldn't have to know `Integer` is preferred.

This just checks for a `Fixnum` type class and coerces it as if it
was `Integer`.