clone default option values when saving
This is needed to ensure the `reset` method works properly.  If you
don't clone and edit the option values without setting them to a new
value, it edits the default directly.  This ensures defaults are always
cloned when resetting to ensure you get the original default value on
each reset.