Fix for setting a proc on a `returned_value` option type
This fixes an edge case when a option with a type that defines a
`returned_value` method (like `NsOptions::Boolean`) is set with
a proc. It wasn't using the `returned_value` method so it would
return an instance of the class instead of the return value of
the `returned_value` method. This makes sure that even if a proc
is used, it will try to use `returned_value`.